# swagger.yaml
openapi: 3.0.0
info:
  title: FinFeed API
  description: Open APIs for FinFeed and its data
  version: 1.0.0
servers:
  - url: "{URL}/api/"
    description: API Route
paths:
...
  /user/:username/posts:
    get:
      tags:
        - user/:username/posts
      summary: Gets all post made by a user
      responses:
        '200':
          description: Successfully received posts
    post:
      tags:
        - user/:username/posts
      summary: Creates a post
      responses:
        '401':
          description: Not authorized to create post on this user
        '200':
          description: Post successfully created
      parameters:
        - name: username
          description: Username of user
          required: true
          schema:
            type: string
  /user/:username/posts/:postID:
    get:
      tags:
        - user/:username/posts
      summary: Gets a specific post from a user
      responses:
        '200':
          description: Post successfully received
      parameters:
        - name: username
          description: Username of user
          required: true
          schema:
            type: string
        - name: postID
          description: ID of the post
          required: true
          schema:
            type: string
    delete:
      tags:
        - user/:username/posts
      summary: Deletes a post
      responses:
        '401':
          description: Not authorized to delete posts
        '200':
          description: Successfully deleted post
      parameters:
        - name: username
          description: Username of user
          required: true
          schema:
            type: string
        - name: postID
          description: ID of the post
          required: true
          schema:
            type: string
  /user/:username/posts/:postID/delete-media:
    delete:
      tags:
        - user/:username/posts
      summary: Deletes a media on a post
      responses:
        '401':
          description: Not authorzied to delete the media on this post
        '200':
          description: Successfully deleted the media off the post
      parameters:
        - name: username
          description: Username of user
          required: true
          schema:
            type: string
        - name: postID
          description: ID of the post
          required: true
          schema:
            type: string
  /user/:username/posts/:postID/edit:
    post:
      tags:
        - user/:username/posts
      summary: Edits a post
      responses:
        '401':
          description: Not authorized to delete this post
        '400':
          description: New edited content missing
        '200':
          description: Successfully edited the post
      parameters:
        - name: username
          description: Username of user
          required: true
          schema:
            type: string
        - name: postID
          description: ID of the post
          required: true
          schema:
            type: string
  /user/:username/posts/:postID/like:
    patch:
      tags:
        - user/:username/posts
      summary: Likes a post
      responses:
        '400':
          description: Liker's username missing
        '401':
          description: Not authorized to like this post
        '200':
          description: Successfully liked post
      parameters:
        - name: username
          description: Username of user
          required: true
          schema:
            type: string
        - name: postID
          description: ID of the post
          required: true
          schema:
            type: string
  /user/:username/posts/:postID/dislike:
    patch:
      tags:
        - user/:username/posts
      summary: Dislikes a post
      responses:
        '400':
          description: Disliker's username missing
        '401':
          description: Not authorized to dislike this post
        '200':
          description: Successfully disliked post
      parameters:
        - name: username
          description: Username of user
          required: true
          schema:
            type: string
        - name: postID
          description: ID of the post
          required: true
          schema:
            type: string
...